weatherApp/requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.0
google-auth==1.27.0
requests==2.25.1
click==6.7
pytz==2017.2
